Fix example app
